added flashcard event setup
