+++ backend/src/routes/flashcardRoutes.js
@@ -8,35 +8,20 @@ class FlashcardRoutes {
   }
 
   initializeRoutes() {
-    // Create Flashcard
+    // Store (Save) Entire Flashcards Dictionary
     this.router.post("/flashcards", async (req, res) => {
       console.log("POST /flashcards");
-      await FlashcardController.createFlashcard(req, res);
+      await FlashcardController.storeFlashcards(req, res);
     });
 
-    // Get All Flashcards
+    // Get Entire Flashcards Dictionary
     this.router.get("/flashcards", async (req, res) => {
       console.log("GET /flashcards");
-      await FlashcardController.getAllFlashcards(req, res);
+      await FlashcardController.getFlashcards(req, res);
     });
 
-    // Get Flashcard By ID
-    this.router.get("/flashcards/:id", async (req, res) => {
-      console.log(`GET /flashcards/${req.params.id}`);
-      await FlashcardController.getFlashcardByID(req, res);
-    });
-
-    // Update Flashcard
-    this.router.put("/flashcards/:id", async (req, res) => {
-      console.log(`PUT /flashcards/${req.params.id}`);
-      await FlashcardController.updateFlashcard(req, res);
-    });
-
-    // Delete Flashcard
-    this.router.delete("/flashcards/:id", async (req, res) => {
-      console.log(`DELETE /flashcards/${req.params.id}`);
-      await FlashcardController.deleteFlashcard(req, res);
-    });
+    // Note: Individual flashcard operations are now managed on the frontend.
+    // The backend only handles the dictionary as a whole, as per the new design.
   }
 
   getRouter() {
